NEW COMPONENT GENERATION SYSTEM! SEE COMMIT DESCRIPTION
Components are no longer created in-code. Instead, a small bit of code
will generate components based on a materail found in the config file.
This means anyone can add a new material at any time, and it will be
fully usable to create guns.
On top of this, it allows my "Brand System" - of which there are two
types; Major brands and Player brands.
Major brands are from the ruins of the old world or whatever server lore
you want to go with. You find components made by these brands, which may
be better or worse than things players can normally craft. They're like
a wildcard.
Player Brands are any major components a player makes (eg receivers),
should they choose to name it. On top of this, Patreon supporters can
get their own brand added to the mod, which will be balanced against
other materials. They simply choose the name and colour, and only they
can craft it. These patreon components will apply to all sub-components
as well (eg firing pins) (It's a cosmetic-only advantage, but a pretty
cool one)